Version 7.1.39 - Added extreme optimization improvements for all player/combat-related tasks - Big rewrite of numerous aspects of the combat system, but no changes in how it should behave (optimizations only) - Players are now able to pick up Elite Coins even when their inventories are full - Elite Coins now follow the player until they get picked up - Bosses now drop more Elite Coins if they are higher level than the player's gear and fewer coins if they are lower level, up to a point where they drop no coins (doesn't affect other loot) - Fixed prestige'd custom rank names not appearing at places - Hunter enchantments now only work in armor slots - Critical enchantments now only work for weapons - Removed VanishNoPacket compatibility - Increased party bonus for natural mob spawning - Natural spawning system no longer scans entire hotbar, only held item for natural mob spawn - Simplified the natural mob spawning logic, tweaked some custom boss dynamic level logic - Fixed duping bug related to trying to wear soulbound armor that belongs to someone else - Fixed NPE related to tracking bosses that no longer exist - Tentative fix for issue related to setting database values - Fixed a bug where regional bosses would never respawn for users of some java version

+++ src/main/java/com/magmaguy/elitemobs/items/ItemTierFinder.java
@@ -93,10 +93,10 @@ public class ItemTierFinder {
 
     }
 
-    private static final double IRON_TIER = CombatSystem.IRON_TIER_LEVEL;
-    private static final double STONE_CHAIN_TIER = CombatSystem.IRON_TIER_LEVEL;
-    private static final double GOLD_WOOD_LEATHER_TIER = CombatSystem.GOLD_WOOD_LEATHER_TIER_LEVEL;
-    private static final double DIAMOND_TIER = CombatSystem.DIAMOND_TIER_LEVEL;
+    private static final int IRON_TIER = CombatSystem.IRON_TIER_LEVEL;
+    private static final int STONE_CHAIN_TIER = CombatSystem.IRON_TIER_LEVEL;
+    private static final int GOLD_WOOD_LEATHER_TIER = CombatSystem.GOLD_WOOD_LEATHER_TIER_LEVEL;
+    private static final int DIAMOND_TIER = CombatSystem.DIAMOND_TIER_LEVEL;
 
     public static double findGenericTier(Material material, HashMap<Enchantment, Integer> enchantmentList) {
 
@@ -132,7 +132,7 @@ public class ItemTierFinder {
 
     }
 
-    public static double findBattleTier(ItemStack itemStack) {
+    public static int findBattleTier(ItemStack itemStack) {
 
         Material material = itemStack.getType();
 
@@ -140,7 +140,7 @@ public class ItemTierFinder {
 
     }
 
-    private static double parseMaterials(Material material, int mainEnchantment, int mainEnchantment2, int mainEnchantment3) {
+    private static int parseMaterials(Material material, int mainEnchantment, int mainEnchantment2, int mainEnchantment3) {
         switch (material) {
             case DIAMOND_SWORD:
             case DIAMOND_AXE:
@@ -218,7 +218,7 @@ public class ItemTierFinder {
         }
     }
 
-    public static double mainHandCombatParser(ItemStack itemStack) {
+    public static int mainHandCombatParser(ItemStack itemStack) {
 
         Material material = itemStack.getType();
 
@@ -226,7 +226,7 @@ public class ItemTierFinder {
 
     }
 
-    private static double parseMainHandMaterials(Material material, int mainEnchantment, int mainEnchantment2, int mainEnchantment3) {
+    private static int parseMainHandMaterials(Material material, int mainEnchantment, int mainEnchantment2, int mainEnchantment3) {
         switch (material) {
             case DIAMOND_SWORD:
             case DIAMOND_AXE:
